@@ -6,38 +6,43 @@ publish_to: 'none'
 version: 0.0.1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  flutter_driver:
+    sdk: flutter
 
-  cupertino_icons: 1.0.0
+  cupertino_icons: ^1.0.3
 
-  http: ^0.12.2
+  http: ^0.13.3
 
-  pedantic: ^1.9.2
+  pedantic: ^1.11.1
 
-  freezed_annotation: ^0.12.0
+  freezed_annotation: ^0.14.3
+  built_value: ^8.1.1
+  built_collection: ^5.1.0
 
-  built_value: ^7.1.0
-  built_collection: ^4.3.2
-
-  redux: ^4.0.0+3
-  flutter_redux: ^0.7.0
-  redux_epics: ^0.14.0
+  redux: ^5.0.0
+  flutter_redux: ^0.8.2
+  redux_epics: ^0.15.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  build_runner: any
-  built_value_generator: ^7.1.0
+  build_runner: ^2.0.6
+  built_value_generator: ^8.1.1
 
-  freezed: ^0.12.0
+  freezed: ^0.14.3
+
+  mocktail: ^0.1.4
 
 dependency_overrides:
-  rxdart: ^0.25.0
+  rxdart: ^0.27.1
+  analyzer: ^2.0.0
+  meta: ^1.7.0
 
 flutter:
   uses-material-design: true
