@@ -17,25 +17,27 @@ class _$AppStateSerializer implements StructuredSerializer<AppState> {
   final String wireName = 'AppState';
 
   @override
-  Iterable<Object> serialize(Serializers serializers, AppState object,
+  Iterable<Object?> serialize(Serializers serializers, AppState object,
       {FullType specifiedType = FullType.unspecified}) {
-    final result = <Object>[
+    final result = <Object?>[
       'searchLocations',
       serializers.serialize(object.searchLocations,
           specifiedType: const FullType(BuiltList, const [const FullType(Location)])),
       'isMetricSystem',
       serializers.serialize(object.isMetricSystem, specifiedType: const FullType(bool)),
     ];
-    if (object.locationForecast != null) {
+    Object? value;
+    value = object.locationForecast;
+    if (value != null) {
       result
         ..add('locationForecast')
-        ..add(serializers.serialize(object.locationForecast, specifiedType: const FullType(LocationForecast)));
+        ..add(serializers.serialize(value, specifiedType: const FullType(LocationForecast)));
     }
     return result;
   }
 
   @override
-  AppState deserialize(Serializers serializers, Iterable<Object> serialized,
+  AppState deserialize(Serializers serializers, Iterable<Object?> serialized,
       {FullType specifiedType = FullType.unspecified}) {
     final result = new AppStateBuilder();
 
@@ -43,18 +45,18 @@ class _$AppStateSerializer implements StructuredSerializer<AppState> {
     while (iterator.moveNext()) {
       final key = iterator.current as String;
       iterator.moveNext();
-      final dynamic value = iterator.current;
+      final Object? value = iterator.current;
       switch (key) {
         case 'searchLocations':
           result.searchLocations.replace(serializers.deserialize(value,
-              specifiedType: const FullType(BuiltList, const [const FullType(Location)])) as BuiltList<Object>);
+              specifiedType: const FullType(BuiltList, const [const FullType(Location)]))! as BuiltList<Object?>);
           break;
         case 'isMetricSystem':
           result.isMetricSystem = serializers.deserialize(value, specifiedType: const FullType(bool)) as bool;
           break;
         case 'locationForecast':
           result.locationForecast.replace(
-              serializers.deserialize(value, specifiedType: const FullType(LocationForecast)) as LocationForecast);
+              serializers.deserialize(value, specifiedType: const FullType(LocationForecast))! as LocationForecast);
           break;
       }
     }
@@ -70,9 +72,9 @@ class _$LocationSerializer implements StructuredSerializer<Location> {
   final String wireName = 'Location';
 
   @override
-  Iterable<Object> serialize(Serializers serializers, Location object,
+  Iterable<Object?> serialize(Serializers serializers, Location object,
       {FullType specifiedType = FullType.unspecified}) {
-    final result = <Object>[
+    final result = <Object?>[
       'woeid',
       serializers.serialize(object.woeid, specifiedType: const FullType(int)),
       'title',
@@ -83,7 +85,7 @@ class _$LocationSerializer implements StructuredSerializer<Location> {
   }
 
   @override
-  Location deserialize(Serializers serializers, Iterable<Object> serialized,
+  Location deserialize(Serializers serializers, Iterable<Object?> serialized,
       {FullType specifiedType = FullType.unspecified}) {
     final result = new LocationBuilder();
 
@@ -91,7 +93,7 @@ class _$LocationSerializer implements StructuredSerializer<Location> {
     while (iterator.moveNext()) {
       final key = iterator.current as String;
       iterator.moveNext();
-      final dynamic value = iterator.current;
+      final Object? value = iterator.current;
       switch (key) {
         case 'woeid':
           result.woeid = serializers.deserialize(value, specifiedType: const FullType(int)) as int;
@@ -113,9 +115,9 @@ class _$LocationForecastSerializer implements StructuredSerializer<LocationForec
   final String wireName = 'LocationForecast';
 
   @override
-  Iterable<Object> serialize(Serializers serializers, LocationForecast object,
+  Iterable<Object?> serialize(Serializers serializers, LocationForecast object,
       {FullType specifiedType = FullType.unspecified}) {
-    final result = <Object>[
+    final result = <Object?>[
       'woeid',
       serializers.serialize(object.woeid, specifiedType: const FullType(int)),
       'title',
@@ -130,7 +132,7 @@ class _$LocationForecastSerializer implements StructuredSerializer<LocationForec
   }
 
   @override
-  LocationForecast deserialize(Serializers serializers, Iterable<Object> serialized,
+  LocationForecast deserialize(Serializers serializers, Iterable<Object?> serialized,
       {FullType specifiedType = FullType.unspecified}) {
     final result = new LocationForecastBuilder();
 
@@ -138,7 +140,7 @@ class _$LocationForecastSerializer implements StructuredSerializer<LocationForec
     while (iterator.moveNext()) {
       final key = iterator.current as String;
       iterator.moveNext();
-      final dynamic value = iterator.current;
+      final Object? value = iterator.current;
       switch (key) {
         case 'woeid':
           result.woeid = serializers.deserialize(value, specifiedType: const FullType(int)) as int;
@@ -165,17 +167,13 @@ class _$AppState extends AppState {
   @override
   final bool isMetricSystem;
   @override
-  final LocationForecast locationForecast;
+  final LocationForecast? locationForecast;
 
-  factory _$AppState([void Function(AppStateBuilder) updates]) => (new AppStateBuilder()..update(updates)).build();
+  factory _$AppState([void Function(AppStateBuilder)? updates]) => (new AppStateBuilder()..update(updates)).build();
 
-  _$AppState._({this.searchLocations, this.isMetricSystem, this.locationForecast}) : super._() {
-    if (searchLocations == null) {
-      throw new BuiltValueNullFieldError('AppState', 'searchLocations');
-    }
-    if (isMetricSystem == null) {
-      throw new BuiltValueNullFieldError('AppState', 'isMetricSystem');
-    }
+  _$AppState._({required this.searchLocations, required this.isMetricSystem, this.locationForecast}) : super._() {
+    BuiltValueNullFieldError.checkNotNull(searchLocations, 'AppState', 'searchLocations');
+    BuiltValueNullFieldError.checkNotNull(isMetricSystem, 'AppState', 'isMetricSystem');
   }
 
   @override
@@ -209,27 +207,28 @@ class _$AppState extends AppState {
 }
 
 class AppStateBuilder implements Builder<AppState, AppStateBuilder> {
-  _$AppState _$v;
+  _$AppState? _$v;
 
-  ListBuilder<Location> _searchLocations;
+  ListBuilder<Location>? _searchLocations;
   ListBuilder<Location> get searchLocations => _$this._searchLocations ??= new ListBuilder<Location>();
-  set searchLocations(ListBuilder<Location> searchLocations) => _$this._searchLocations = searchLocations;
+  set searchLocations(ListBuilder<Location>? searchLocations) => _$this._searchLocations = searchLocations;
 
-  bool _isMetricSystem;
-  bool get isMetricSystem => _$this._isMetricSystem;
-  set isMetricSystem(bool isMetricSystem) => _$this._isMetricSystem = isMetricSystem;
+  bool? _isMetricSystem;
+  bool? get isMetricSystem => _$this._isMetricSystem;
+  set isMetricSystem(bool? isMetricSystem) => _$this._isMetricSystem = isMetricSystem;
 
-  LocationForecastBuilder _locationForecast;
+  LocationForecastBuilder? _locationForecast;
   LocationForecastBuilder get locationForecast => _$this._locationForecast ??= new LocationForecastBuilder();
-  set locationForecast(LocationForecastBuilder locationForecast) => _$this._locationForecast = locationForecast;
+  set locationForecast(LocationForecastBuilder? locationForecast) => _$this._locationForecast = locationForecast;
 
   AppStateBuilder();
 
   AppStateBuilder get _$this {
-    if (_$v != null) {
-      _searchLocations = _$v.searchLocations?.toBuilder();
-      _isMetricSystem = _$v.isMetricSystem;
-      _locationForecast = _$v.locationForecast?.toBuilder();
+    final $v = _$v;
+    if ($v != null) {
+      _searchLocations = $v.searchLocations.toBuilder();
+      _isMetricSystem = $v.isMetricSystem;
+      _locationForecast = $v.locationForecast?.toBuilder();
       _$v = null;
     }
     return this;
@@ -237,14 +236,12 @@ class AppStateBuilder implements Builder<AppState, AppStateBuilder> {
 
   @override
   void replace(AppState other) {
-    if (other == null) {
-      throw new ArgumentError.notNull('other');
-    }
+    ArgumentError.checkNotNull(other, 'other');
     _$v = other as _$AppState;
   }
 
   @override
-  void update(void Function(AppStateBuilder) updates) {
+  void update(void Function(AppStateBuilder)? updates) {
     if (updates != null) updates(this);
   }
 
@@ -255,10 +252,10 @@ class AppStateBuilder implements Builder<AppState, AppStateBuilder> {
       _$result = _$v ??
           new _$AppState._(
               searchLocations: searchLocations.build(),
-              isMetricSystem: isMetricSystem,
+              isMetricSystem: BuiltValueNullFieldError.checkNotNull(isMetricSystem, 'AppState', 'isMetricSystem'),
               locationForecast: _locationForecast?.build());
     } catch (_) {
-      String _$failedField;
+      late String _$failedField;
       try {
         _$failedField = 'searchLocations';
         searchLocations.build();
@@ -281,15 +278,11 @@ class _$Location extends Location {
   @override
   final String title;
 
-  factory _$Location([void Function(LocationBuilder) updates]) => (new LocationBuilder()..update(updates)).build();
+  factory _$Location([void Function(LocationBuilder)? updates]) => (new LocationBuilder()..update(updates)).build();
 
-  _$Location._({this.woeid, this.title}) : super._() {
-    if (woeid == null) {
-      throw new BuiltValueNullFieldError('Location', 'woeid');
-    }
-    if (title == null) {
-      throw new BuiltValueNullFieldError('Location', 'title');
-    }
+  _$Location._({required this.woeid, required this.title}) : super._() {
+    BuiltValueNullFieldError.checkNotNull(woeid, 'Location', 'woeid');
+    BuiltValueNullFieldError.checkNotNull(title, 'Location', 'title');
   }
 
   @override
@@ -316,22 +309,23 @@ class _$Location extends Location {
 }
 
 class LocationBuilder implements Builder<Location, LocationBuilder> {
-  _$Location _$v;
+  _$Location? _$v;
 
-  int _woeid;
-  int get woeid => _$this._woeid;
-  set woeid(int woeid) => _$this._woeid = woeid;
+  int? _woeid;
+  int? get woeid => _$this._woeid;
+  set woeid(int? woeid) => _$this._woeid = woeid;
 
-  String _title;
-  String get title => _$this._title;
-  set title(String title) => _$this._title = title;
+  String? _title;
+  String? get title => _$this._title;
+  set title(String? title) => _$this._title = title;
 
   LocationBuilder();
 
   LocationBuilder get _$this {
-    if (_$v != null) {
-      _woeid = _$v.woeid;
-      _title = _$v.title;
+    final $v = _$v;
+    if ($v != null) {
+      _woeid = $v.woeid;
+      _title = $v.title;
       _$v = null;
     }
     return this;
@@ -339,20 +333,21 @@ class LocationBuilder implements Builder<Location, LocationBuilder> {
 
   @override
   void replace(Location other) {
-    if (other == null) {
-      throw new ArgumentError.notNull('other');
-    }
+    ArgumentError.checkNotNull(other, 'other');
     _$v = other as _$Location;
   }
 
   @override
-  void update(void Function(LocationBuilder) updates) {
+  void update(void Function(LocationBuilder)? updates) {
     if (updates != null) updates(this);
   }
 
   @override
   _$Location build() {
-    final _$result = _$v ?? new _$Location._(woeid: woeid, title: title);
+    final _$result = _$v ??
+        new _$Location._(
+            woeid: BuiltValueNullFieldError.checkNotNull(woeid, 'Location', 'woeid'),
+            title: BuiltValueNullFieldError.checkNotNull(title, 'Location', 'title'));
     replace(_$result);
     return _$result;
   }
@@ -368,22 +363,15 @@ class _$LocationForecast extends LocationForecast {
   @override
   final double temperature;
 
-  factory _$LocationForecast([void Function(LocationForecastBuilder) updates]) =>
+  factory _$LocationForecast([void Function(LocationForecastBuilder)? updates]) =>
       (new LocationForecastBuilder()..update(updates)).build();
 
-  _$LocationForecast._({this.woeid, this.title, this.weather, this.temperature}) : super._() {
-    if (woeid == null) {
-      throw new BuiltValueNullFieldError('LocationForecast', 'woeid');
-    }
-    if (title == null) {
-      throw new BuiltValueNullFieldError('LocationForecast', 'title');
-    }
-    if (weather == null) {
-      throw new BuiltValueNullFieldError('LocationForecast', 'weather');
-    }
-    if (temperature == null) {
-      throw new BuiltValueNullFieldError('LocationForecast', 'temperature');
-    }
+  _$LocationForecast._({required this.woeid, required this.title, required this.weather, required this.temperature})
+      : super._() {
+    BuiltValueNullFieldError.checkNotNull(woeid, 'LocationForecast', 'woeid');
+    BuiltValueNullFieldError.checkNotNull(title, 'LocationForecast', 'title');
+    BuiltValueNullFieldError.checkNotNull(weather, 'LocationForecast', 'weather');
+    BuiltValueNullFieldError.checkNotNull(temperature, 'LocationForecast', 'temperature');
   }
 
   @override
@@ -419,32 +407,33 @@ class _$LocationForecast extends LocationForecast {
 }
 
 class LocationForecastBuilder implements Builder<LocationForecast, LocationForecastBuilder> {
-  _$LocationForecast _$v;
+  _$LocationForecast? _$v;
 
-  int _woeid;
-  int get woeid => _$this._woeid;
-  set woeid(int woeid) => _$this._woeid = woeid;
+  int? _woeid;
+  int? get woeid => _$this._woeid;
+  set woeid(int? woeid) => _$this._woeid = woeid;
 
-  String _title;
-  String get title => _$this._title;
-  set title(String title) => _$this._title = title;
+  String? _title;
+  String? get title => _$this._title;
+  set title(String? title) => _$this._title = title;
 
-  String _weather;
-  String get weather => _$this._weather;
-  set weather(String weather) => _$this._weather = weather;
+  String? _weather;
+  String? get weather => _$this._weather;
+  set weather(String? weather) => _$this._weather = weather;
 
-  double _temperature;
-  double get temperature => _$this._temperature;
-  set temperature(double temperature) => _$this._temperature = temperature;
+  double? _temperature;
+  double? get temperature => _$this._temperature;
+  set temperature(double? temperature) => _$this._temperature = temperature;
 
   LocationForecastBuilder();
 
   LocationForecastBuilder get _$this {
-    if (_$v != null) {
-      _woeid = _$v.woeid;
-      _title = _$v.title;
-      _weather = _$v.weather;
-      _temperature = _$v.temperature;
+    final $v = _$v;
+    if ($v != null) {
+      _woeid = $v.woeid;
+      _title = $v.title;
+      _weather = $v.weather;
+      _temperature = $v.temperature;
       _$v = null;
     }
     return this;
@@ -452,24 +441,26 @@ class LocationForecastBuilder implements Builder<LocationForecast, LocationForec
 
   @override
   void replace(LocationForecast other) {
-    if (other == null) {
-      throw new ArgumentError.notNull('other');
-    }
+    ArgumentError.checkNotNull(other, 'other');
     _$v = other as _$LocationForecast;
   }
 
   @override
-  void update(void Function(LocationForecastBuilder) updates) {
+  void update(void Function(LocationForecastBuilder)? updates) {
     if (updates != null) updates(this);
   }
 
   @override
   _$LocationForecast build() {
-    final _$result =
-        _$v ?? new _$LocationForecast._(woeid: woeid, title: title, weather: weather, temperature: temperature);
+    final _$result = _$v ??
+        new _$LocationForecast._(
+            woeid: BuiltValueNullFieldError.checkNotNull(woeid, 'LocationForecast', 'woeid'),
+            title: BuiltValueNullFieldError.checkNotNull(title, 'LocationForecast', 'title'),
+            weather: BuiltValueNullFieldError.checkNotNull(weather, 'LocationForecast', 'weather'),
+            temperature: BuiltValueNullFieldError.checkNotNull(temperature, 'LocationForecast', 'temperature'));
     replace(_$result);
     return _$result;
   }
 }
 
-// ignore_for_file: always_put_control_body_on_new_line,always_specify_types,annotate_overrides,avoid_annotating_with_dynamic,avoid_as,avoid_catches_without_on_clauses,avoid_returning_this,lines_longer_than_80_chars,omit_local_variable_types,prefer_expression_function_bodies,sort_constructors_first,test_types_in_equals,unnecessary_const,unnecessary_new
+// ignore_for_file: always_put_control_body_on_new_line,always_specify_types,annotate_overrides,avoid_annotating_with_dynamic,avoid_as,avoid_catches_without_on_clauses,avoid_returning_this,deprecated_member_use_from_same_package,lines_longer_than_80_chars,omit_local_variable_types,prefer_expression_function_bodies,sort_constructors_first,test_types_in_equals,unnecessary_const,unnecessary_new
